🎨 chore: 중복 색상, 미사용 모듈 제거

+++ finafanAndroid/app/src/main/java/com/a702/finafan/presentation/funding/FundingProgressBar.kt
@@ -23,8 +23,8 @@ import com.a702.finafan.common.ui.theme.Pretendard
 
 @Composable
 fun FundingProgressBar(
-    currentAmount: Int,
-    goalAmount: Int,
+    currentAmount: Long,
+    goalAmount: Long,
     gradientColors: List<Color>
 ) {
